ingredients authorization

+++ RecipesAPI/Controllers/IngredientsController.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using RecipesAPI.Auth.Model;
 using RecipesAPI.Data.Dtos.Ingredients;
 using RecipesAPI.Data.Entities;
 using RecipesAPI.Data.Repositories;
@@ -18,12 +20,14 @@ namespace RecipesAPI.Controllers
         private readonly IIngredientsRepository _ingredientsRepository;
         private readonly IMapper _mapper;
         private readonly IRecipesRepository _recipesRepository;
+        private readonly IAuthorizationService _authorizationService;
 
-        public IngredientsController(IIngredientsRepository ingredientsRepository, IMapper mapper, IRecipesRepository recipesRepository)
+        public IngredientsController(IIngredientsRepository ingredientsRepository, IMapper mapper, IRecipesRepository recipesRepository, IAuthorizationService authorizationService)
         {
             _ingredientsRepository = ingredientsRepository;
             _mapper = mapper;
             _recipesRepository = recipesRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -56,12 +60,19 @@ namespace RecipesAPI.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = RestUserRoles.SimpleUser)]
         public async Task<ActionResult<IngredientDto>> PostAsync(int recipeId, CreateIngredientDto ingredientDto)
         {
             var recipe = await _recipesRepository.GetAsync(recipeId);
             if (recipe == null)
             {
                 return NotFound($"Recipe with id '{recipeId}' not found.");
+            }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, recipe, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
             }
 
             var ingredient = _mapper.Map<Ingredient>(ingredientDto);
@@ -75,6 +86,7 @@ namespace RecipesAPI.Controllers
         }
 
         [HttpPut("{ingredientId}")]
+        [Authorize(Roles = "" + RestUserRoles.SimpleUser + "," + RestUserRoles.Admin + "")]
         public async Task<ActionResult<IngredientDto>> PutAsync(int recipeId, int ingredientId, UpdateIngredientDto ingredientDto)
         {
             var recipe = await _recipesRepository.GetAsync(recipeId);
@@ -83,11 +95,19 @@ namespace RecipesAPI.Controllers
                 return NotFound($"Recipe with id '{recipeId}' not found.");
             }
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, recipe, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             var oldIngredient = await _ingredientsRepository.GetAsync(recipeId, ingredientId);
             if (oldIngredient == null)
             {
                 return NotFound($"Ingredient with id '{ingredientId}' not found.");
             }
+
+            //TODO: ingredient authorization ?
 
             _mapper.Map(ingredientDto, oldIngredient);
 
@@ -97,6 +117,7 @@ namespace RecipesAPI.Controllers
         }
 
         [HttpDelete("{ingredientId}")]
+        [Authorize(Roles = "" + RestUserRoles.SimpleUser + "," + RestUserRoles.Admin + "")]
         public async Task<ActionResult> DeleteAsync(int recipeId, int ingredientId)
         {
             var recipe = await _recipesRepository.GetAsync(recipeId);
@@ -105,11 +126,19 @@ namespace RecipesAPI.Controllers
                 return NotFound($"Recipe with id '{recipeId}' not found.");
             }
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, recipe, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             var ingredient = await _ingredientsRepository.GetAsync(recipeId, ingredientId);
             if (ingredient == null)
             {
                 return NotFound($"Ingredient with id '{ingredientId}' not found.");
             }
+
+            // TODO: ingredient authorization ?
 
             await _ingredientsRepository.DeleteAsync(ingredient);
